+++ common/index.ts
@@ -1,16 +1,6 @@
-export * from './hooks/UseSpacingStylesProps';
-export * from './Box/BoxBaseProps';
-export * from './Button/ButtonBaseProps';
-export * from './Text/TextBaseProps';
-export * from './Text/Typography';
-export * from './types';
-
-export * from './SystemProvider';
+export * from './palette/constants';
 export * from './palette/usePaletteConfig';
 export * from './scale/useScale';
 export * from './spectrum/useSpectrum';
-
-export * from './palette/constants';
-export * from './palette/types';
-export * from './scale/types';
-export * from './spectrum/types';
+export * from './SystemProvider';
+export * from './types';