+++ common/types/ElevationLevels.ts
@@ -1,0 +1,1 @@
+export type ElevationLevels = 1 | 2;