+++ common/types/ForwardedRef.ts
@@ -1,0 +1,3 @@
+import type { MutableRefObject } from 'react';
+
+export type ForwardedRef<T> = ((instance: T | null) => void) | MutableRefObject<T | null> | null;