+++ common/types/OffsetBaseProps.ts
@@ -1,0 +1,3 @@
+import type { UseSpacingStylesProps } from './UseSpacingStylesProps';
+
+export type OffsetBaseProps = Omit<UseSpacingStylesProps, 'isInverted'>;