+++ common/types/SpacingProps.ts
@@ -1,0 +1,11 @@
+import type { SpacingScale } from './SpacingScale';
+
+export interface SpacingProps {
+  spacing?: SpacingScale;
+  spacingBottom?: SpacingScale;
+  spacingEnd?: SpacingScale;
+  spacingHorizontal?: SpacingScale;
+  spacingStart?: SpacingScale;
+  spacingTop?: SpacingScale;
+  spacingVertical?: SpacingScale;
+}